step3/fix02
