+++ src/main/java/org/homework/Status.java
@@ -1,0 +1,5 @@
+package org.homework;
+
+public enum Status {
+    완료, 미완료
+}
