+++ src/main/java/org/homework/TodoRepository.java
@@ -1,9 +1,7 @@
 package org.homework;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
@@ -28,32 +26,12 @@ public class TodoRepository {
     public Optional<Todo> changeStatus(int id) {
         return Optional.ofNullable(todoMap.get(id))
                 .map(todo -> {
-                    todo.setStatus(!todo.isStatus());
+                    todo.setStatus(todo.getStatus() == Status.완료 ? Status.미완료 : Status.완료);
                     return todo;
                 });
     }
 
-    public List<Todo> findToDoList() {
-        List<Todo> todoList = new ArrayList<>();
-        todoMap.values().stream()
-                .filter(this::is7DaysLeft)
-                .sorted((o1, o2) -> o2.getDeadLine().compareTo(o1.getDeadLine()))
-                .forEach(todoList::add);
-        return todoList;
-    }
-
-    public List<Todo> findKeyword(String keyword) {
-        List<Todo> todoList = new ArrayList<>();
-        todoMap.values().stream()
-                .filter(todo -> todo.getContents().contains(keyword))
-                .sorted((o1, o2) -> o2.getDeadLine().compareTo(o1.getDeadLine()))
-                .forEach(todoList::add);
-        return todoList;
-    }
-
-    private boolean is7DaysLeft(Todo todo) {
-        LocalDate today = LocalDate.now();
-        LocalDate deadline = todo.getDeadLine();
-        return !deadline.isBefore(today) && !deadline.isAfter(today.plusDays(7));
+    public Map<Integer, Todo> getTodoMap() {
+        return todoMap;
     }
 }
