+++ src/main/java/org/homework/TodoService.java
@@ -2,7 +2,7 @@ package org.homework;
 
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Optional;
+import java.util.stream.Collectors;
 
 public class TodoService {
     private TodoRepository todoRepository = new TodoRepository();
@@ -27,10 +27,22 @@ public class TodoService {
     }
 
     public List<Todo> findToDoList() {
-        return todoRepository.findToDoList();
+        return todoRepository.getTodoMap().values().stream()
+                .filter(this::is7DaysLeft)
+                .sorted((o1, o2) -> o2.getDeadLine().compareTo(o1.getDeadLine()))
+                .collect(Collectors.toList());
     }
 
     public List<Todo> findKeyword(String keyword) {
-        return todoRepository.findKeyword(keyword);
+        return todoRepository.getTodoMap().values().stream()
+                .filter(todo -> todo.getContents().contains(keyword))
+                .sorted((o1, o2) -> o2.getDeadLine().compareTo(o1.getDeadLine()))
+                .collect(Collectors.toList());
+    }
+
+    private boolean is7DaysLeft(Todo todo) {
+        LocalDate today = LocalDate.now();
+        LocalDate deadline = todo.getDeadLine();
+        return !deadline.isBefore(today) && !deadline.isAfter(today.plusDays(7));
     }
 }
